requirements.txt
ccxt>=4.2.0
pandas>=2.0.0
TA-Lib>=0.4.24
python-telegram-bot>=20.0.0
python-dotenv>=1.0.0
